reorg dash page components.py functions into page-specific service modules
